restructure builtin_msgs support to fix ros2/examples#2
