+++ rosidl_typesupport_opensplice_cpp/include/rosidl_typesupport_opensplice_cpp/Time_TypeSupport.h
@@ -15,7 +15,7 @@
 #ifndef __rosidl_typesupport_openslice_cpp__Time_TypeSupport__h__
 #define __rosidl_typesupport_openslice_cpp__Time_TypeSupport__h__
 
-#include <iostream>
+#include <ccpp_dds_dcps.h>
 
 #include "builtin_msgs/Time_Struct.h"
 
@@ -24,17 +24,9 @@ namespace builtin_msgs
 namespace type_support
 {
 
-void convert_ros_message_to_dds(const builtin_msgs::Time & ros_message, DDS::Time_t & dds_message)
-{
-  dds_message.sec = ros_message.sec;
-  dds_message.nanosec = ros_message.nanosec;
-}
+void convert_ros_message_to_dds(const builtin_msgs::Time & ros_message, DDS::Time_t & dds_message);
 
-void convert_dds_message_to_ros(const DDS::Time_t & dds_message, builtin_msgs::Time & ros_message)
-{
-  ros_message.sec = dds_message.sec;
-  ros_message.nanosec = dds_message.nanosec;
-}
+void convert_dds_message_to_ros(const DDS::Time_t & dds_message, builtin_msgs::Time & ros_message);
 
 }  // namespace type_support
 }  // namespace builtin_msgs
