+++ rosidl_typesupport_opensplice_cpp/resource/Time_TypeSupport.cpp
@@ -12,27 +12,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef __rosidl_typesupport_openslice_cpp__Duration_TypeSupport__h__
-#define __rosidl_typesupport_openslice_cpp__Duration_TypeSupport__h__
-
-#include <ccpp_dds_dcps.h>
-
-#include "builtin_msgs/Duration_Struct.h"
+#include "rosidl_typesupport_opensplice_cpp/Time_TypeSupport.h"
 
 namespace builtin_msgs
 {
 namespace type_support
 {
 
-void convert_ros_message_to_dds(
-  const builtin_msgs::Duration & ros_message,
-  DDS::Duration_t & dds_message);
+void convert_ros_message_to_dds(const builtin_msgs::Time & ros_message, DDS::Time_t & dds_message)
+{
+  dds_message.sec = ros_message.sec;
+  dds_message.nanosec = ros_message.nanosec;
+}
 
-void convert_dds_message_to_ros(
-  const DDS::Duration_t & dds_message,
-  builtin_msgs::Duration & ros_message);
+void convert_dds_message_to_ros(const DDS::Time_t & dds_message, builtin_msgs::Time & ros_message)
+{
+  ros_message.sec = dds_message.sec;
+  ros_message.nanosec = dds_message.nanosec;
+}
 
 }  // namespace type_support
 }  // namespace builtin_msgs
-
-#endif  // __rosidl_typesupport_openslice_cpp__Duration_TypeSupport__h__
